<fix>[zstackctl]: upgrade python3
upgrade zstack-ctl zstack-cli
kvmagent cephagent consoleproxy
to python 3.11

upgrade most of dependency

fix most of warning.

Resolves: ZSTAC-73161

+++ iscsifilesystemagent/iscsifilesystemagent/plugins/btrfs.py
@@ -103,7 +103,7 @@ class BtrfsPlugin(plugin.Plugin):
         if not avail:
             raise Exception('cannot find available capacity, "df" output":\n %s' % out)
 
-        return total, long(avail) * 1024
+        return total, int(avail) * 1024
 
     @iscsiagent.replyerror
     def init(self, req):
